ChunkyMonkey.js added
This is currently not operational, but is an initial implementation that is meant to take an array, split it up into chunks the size of value of the size variable. Handling of instances wherein the array does not equally divide into chunks of size-sized sub-arrays should place the remainder into its own sub-array at the end of the matrix.